split test_heartbeat_detection to match module structure
